FEAT: Turns out Terraform hates unknown keys
You often might want to create permissions at the
same time as the resource itself. This is an
issue in the current version of the module
as Terraform will claim it doesn't know
the number of resources to create. We fix
this by reverting to using an old-fashioned
count instead of for_each. As the for_each
key was already an index since the last
feature anyway this should not pose any
ergonomic issues when using it.